Update .gitignore to exclude testem.log, /typings, and .vs
The .gitignore file has been updated to include the testem.log file,
which means this file will now be ignored by Git and not tracked in
the repository. The /typings directory has been added to the
.gitignore file, indicating that this directory will also be ignored
by Git. The .vs directory has been added to the .gitignore file,
which means that Visual Studio specific files will be ignored by Git.